1. Modified error capture for missing files, existing folders and existing files. 2. Will not recreate existing folders. 3. Will not move existing file. 4. Counts number of files found, folders created and files moved.

+++ main_window.py
@@ -26,6 +26,8 @@ class MainWindow:
 
         self.files_to_process = NONE
         self.n_files_to_process = 0
+        self.n_files_moved = 0
+        self.n_folders_created = 0
 
         # GUI Frames
         self.frame_root_title = Frame(root, highlightthickness=0)
@@ -48,23 +50,28 @@ class MainWindow:
 
         # Entries
         self.entry_files_location_entry = StringVar()
-        self.entry_files_location_entry.trace("w", lambda name, index, mode, entry_files_location_entry=self.entry_files_location_entry: self.entry_update_files_location())
-        self.entry_files_location = Entry(self.frame_root_session, width=80, textvariable=self.entry_files_location_entry)
+        self.entry_files_location_entry.trace("w", lambda name, index, mode,
+                                                          entry_files_location_entry=self.entry_files_location_entry: self.entry_update_files_location())
+        self.entry_files_location = Entry(self.frame_root_session, width=80,
+                                          textvariable=self.entry_files_location_entry)
         self.entry_files_location.insert(END, os.path.normcase(FOLDER_PATH))
 
         # Buttons
-        self.button_files_location = Button(self.frame_root_session, text="Files Location", command=lambda: self.choose_folder(), pady=0, width=10, fg='brown')
-        self.button_start = Button(self.frame_root_session, text="Start", fg='green', command=self.start_process, pady=0, width=10)
-        self.button_exit = Button(self.frame_root_session, text="Exit", fg='red', command=self.quit_program, pady=0, width=10)
+        self.button_files_location = Button(self.frame_root_session, text="Files Location",
+                                            command=lambda: self.choose_folder(), pady=0, width=10, fg='brown')
+        self.button_start = Button(self.frame_root_session, text="Start", fg='green', command=self.start_process,
+                                   pady=0, width=10)
+        self.button_exit = Button(self.frame_root_session, text="Exit", fg='red', command=self.quit_program, pady=0,
+                                  width=10)
         self.button_open_folder = Button(self.frame_root_session, text="Open Folder",
                                          command=lambda: self.open_folder(),
                                          pady=0, width=10)
         # Grids
-        self.entry_files_location.grid      (row=1, column=1, sticky=W, padx=(5, 0))
-        self.button_files_location.grid     (row=1, column=0, sticky=NE)
-        self.button_start.grid              (row=2, column=0, sticky=NE)
-        self.button_open_folder.grid        (row=2, column=1, sticky=NW)
-        self.button_exit.grid               (row=2, column=1, sticky=NW, padx=(80, 0))
+        self.entry_files_location.grid(row=1, column=1, sticky=W, padx=(5, 0))
+        self.button_files_location.grid(row=1, column=0, sticky=NE)
+        self.button_start.grid(row=2, column=0, sticky=NE)
+        self.button_open_folder.grid(row=2, column=1, sticky=NW)
+        self.button_exit.grid(row=2, column=1, sticky=NW, padx=(80, 0))
 
         # END OF FRAME #######################################################
 
@@ -87,54 +94,86 @@ class MainWindow:
         sys.exit()
 
     def start_process(self):
-        self.scan_files_in_folder(self.user_entry.folder_path)
-        self.create_folders_and_move_files(self.files_to_process, self.user_entry.folder_path)
+        self.n_files_to_process = 0
+        self.n_files_moved = 0
+        self.n_folders_created = 0
+
+        # Scan Folder
+        result = self.scan_files_in_folder(self.user_entry.folder_path)
+
+        # No Files
+        if result == 0:
+            message = "No Files to Process: " + str(self.n_files_to_process) + '\n'
+            colour = 'red'
+            self.session_log.write_textbox(message, colour)
+            return
+
+        # No Folder
+        elif result == -1:
+            message = "Folder path does not exist .. Process will terminate" + '\n'
+            colour = 'brown'
+            self.session_log.write_textbox(message, colour)
+            print(message)
+            return
+
+        # Files Detected
+        else:
+            self.create_folders_and_move_files(self.files_to_process, self.user_entry.folder_path)
+
+            if self.n_files_to_process == self.n_folders_created == self.n_files_moved:
+                message = "Number of Folders Created: " + str(self.n_folders_created) + '\n'
+                colour = 'green'
+                self.session_log.write_textbox(message, colour)
+                message = "Number of Files Moved: " + str(self.n_files_moved) + '\n'
+                colour = 'green'
+                self.session_log.write_textbox(message, colour)
+            else:
+                message = "WARNING: MISMATCH" + '\n'
+                colour = 'red'
+                self.session_log.write_textbox(message, colour)
+                message = "Number of Folders Created: " + str(self.n_folders_created) + '\n'
+                colour = 'red'
+                self.session_log.write_textbox(message, colour)
+                message = "Number of Files Moved: " + str(self.n_files_moved) + '\n'
+                colour = 'red'
+                self.session_log.write_textbox(message, colour)
 
     def scan_files_in_folder(self, folder_path):
-        # Check if folder path exists
-        # if os.path.exists(folder_path):
-        try:
-            message = "Scanning Folder" + '\n'
-            colour = "black"
-            self.session_log.write_textbox(message, colour)
+        message = "Scanning Folder" + '\n'
+        colour = "black"
+        self.session_log.write_textbox(message, colour)
 
+        # Check if folder path exists
+        if os.path.exists(folder_path):
             # Get a list of files in the folder
             self.files_to_process = [f for f in os.listdir(folder_path) if os.path.isfile(os.path.join(folder_path, f))]
             self.n_files_to_process = len(self.files_to_process)
 
-            message = "Number of Files to Process: " + str(self.n_files_to_process) + '\n'
-            colour = 'blue'
-            self.session_log.write_textbox(message, colour)
+            if self.n_files_to_process < 1:
+                return 0
+
+            else:
+                message = "Number of Files to Process: " + str(self.n_files_to_process) + '\n'
+                colour = 'blue'
+                self.session_log.write_textbox(message, colour)
 
             message = 'Files List:' + '\n'
             colour = 'black'
             self.session_log.write_textbox(message, colour)
 
-            print("Files in folder:")
-            for file in self.files_to_process:
-                print(file)
-                message = '\t' + str(file) + '\n'
-                colour = 'black'
-                self.session_log.write_textbox(message, colour)
-            return self.files_to_process
+            return 1
 
-        # else:
-        except:
-            message = "Folder path does not exist .. Process will terminate" + '\n'
-            colour = 'red'
-            self.session_log.write_textbox(message, colour)
-            print(message)
-            raise SystemExit
+        else:
+            return -1
 
     def create_folders_and_move_files(self, filenames_list, destination_folder):
-
         # Check if the destination folder exists
         if not os.path.exists(destination_folder):
-            message = "Destination folder does not exist" + '\n'
-            colour = "red"
+            message = "Folder path does not exist .. Process will terminate" + '\n'
+            colour = 'brown'
             self.session_log.write_textbox(message, colour)
             print(message)
-            return
+            return -1
 
         for filename in filenames_list:
             # Create a folder using the filename in the destination folder
@@ -144,20 +183,38 @@ class MainWindow:
             # Create the folder
             try:
                 os.makedirs(folder_path)
+                self.n_folders_created += 1
                 print(f"Folder created: {folder_path}")
             except FileExistsError:
-                print(f"Folder already exists: {folder_path}")
+                message = f"Folder already exists: {folder_path}" + '\n'
+                colour = 'brown'
+                self.session_log.write_textbox(message, colour)
+                print(message)
 
             # Move the file to the created folder
-            file_path = os.path.join(destination_folder, filename)
-            new_file_path = os.path.join(folder_path, filename)
             try:
-                shutil.move(file_path, new_file_path)
-                print(f"File moved: {filename} -> {new_file_path}")
+                file_path = os.path.join(destination_folder, filename)
+                new_file_path = os.path.join(folder_path, filename)
+
+                if not os.path.isfile(new_file_path):
+                    shutil.move(file_path, new_file_path)
+                    self.n_files_moved += 1
+                    message = str(filename) + "  >>>  " + str(os.path.dirname(new_file_path)) + '\n'
+                    colour = 'black'
+                    self.session_log.write_textbox(message, colour)
+                    print(message)
+
+                else:
+                    message = f"File {file_path} Exists and Will Be Ignored" + '\n'
+                    colour = 'red'
+                    self.session_log.write_textbox(message, colour)
+                    print(message)
+
             except FileNotFoundError:
-                print(f"File not found: {filename}")
-            except shutil.Error:
-                print(f"File already exists in the folder: {filename}")
+                message = f"File not found: {filename}" + '\n'
+                colour = 'red'
+                self.session_log.write_textbox(message, colour)
+                print(message)
 
     def open_folder(self):
         temp_path = os.path.realpath(self.user_entry.folder_path)
